+++ src/includes/classes/SCore/Base/Facades.php
@@ -45,6 +45,7 @@ abstract class Facades
     use Traits\Facades\TermQueries;
     use Traits\Facades\Theme;
     use Traits\Facades\Transients;
+    use Traits\Facades\TransientShortlink;
     use Traits\Facades\Uninstaller;
     use Traits\Facades\UrlToPostId;
     use Traits\Facades\UserQueries;